fix: relocated project Studio and Class in JCF: List
